Replace context hook with direct call
