+++ kedro_kubeflow/context_helper.py
@@ -1,10 +1,37 @@
+import os
 from functools import lru_cache
 from pathlib import Path
+from typing import Dict, Iterable
 
 from kedro import __version__ as kedro_version
+from kedro.config import TemplatedConfigLoader
 from semver import VersionInfo
 
 from .config import PluginConfig
+
+
+class EnvTemplatedConfigLoader(TemplatedConfigLoader):
+    """Config loader that can substitute $(commit_id) and $(branch_name)
+    placeholders with information taken from env variables."""
+
+    VAR_PREFIX = "KEDRO_CONFIG_"
+    # defaults provided so default variables ${commit_id|dirty} work for some entries
+    ENV_DEFAULTS = {"commit_id": None, "branch_name": None}
+
+    def __init__(self, conf_paths: Iterable[str]):
+        super().__init__(conf_paths, globals_dict=self.read_env())
+
+    def read_env(self) -> Dict:
+        config = EnvTemplatedConfigLoader.ENV_DEFAULTS.copy()
+        overrides = dict(
+            [
+                (k.replace(EnvTemplatedConfigLoader.VAR_PREFIX, "").lower(), v)
+                for k, v in os.environ.copy().items()
+                if k.startswith(EnvTemplatedConfigLoader.VAR_PREFIX)
+            ]
+        )
+        config.update(**overrides)
+        return config
 
 
 class ContextHelper(object):
@@ -33,7 +60,9 @@ class ContextHelper(object):
     @property
     @lru_cache()
     def config(self) -> PluginConfig:
-        raw = self.context.config_loader.get(self.CONFIG_FILE_PATTERN)
+        raw = EnvTemplatedConfigLoader(
+            self.context.config_loader.conf_paths
+        ).get(self.CONFIG_FILE_PATTERN)
         return PluginConfig(raw)
 
     @property
